lab7 done
finish virtual file system with /dev/uart and /dev/framebuffer.

+++ lib/mbox.c
@@ -5,7 +5,7 @@
 volatile unsigned int  __attribute__((aligned(16))) mbox[36];
 
 int _mbox_call(unsigned char ch) {
-    unsigned int r = (((unsigned int)((unsigned long)&mbox) & ~0xF) | (ch & 0xF));
+    unsigned long r = (((unsigned long)((unsigned long)&mbox) & ~0xF) | (ch & 0xF));
     // wait for ready 
     do {
         asm volatile("nop");
